requirements.txt
###### Core utilities
# Load your OpenAI API key from .env
python-dotenv
# Show progress bars when evaluating many prompts
tqdm
# For metrics, tabular results, data analysis
numpy
pandas

###### NLP / Evaluation
# For BLEU and text metrics
nltk

###### Model APIs
# To call GPT models
openai
requests
pytest
huggingface_hub